pyproject.toml
[project]
name = "mcp-python-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "fastmcp[http]>=2.5.1",
    "psycopg2>=2.9.10",
    "sqlalchemy>=2.0.41",
]
